типизирует
