+++ src/components/ProtectedRoute/ProtectedRoute.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { useSelector } from 'react-redux';
 import { Navigate, useLocation } from 'react-router-dom';
 
@@ -5,7 +6,10 @@ const Protected = ({ onlyUnAuth = false, component }) => {
 	// isAuthChecked это флаг, показывающий что проверка токена произведена
 	// при этом результат этой проверки не имеет значения, важно только,
 	// что сам факт проверки имел место.
+	//@ts-ignore
 	const isAuthChecked = useSelector((store) => store.loginReducer.isAuthChecked);
+
+	//@ts-ignore
 	const user = useSelector((store) => store.loginReducer.name);
 
 	const location = useLocation();
